Renamed inp to input, added test for it, fixed restart logging.
node.restart() logging was previously broken: -l option was not supplied, so its
stdout/err wasn't redirected -- this is option of pg_ctl, not of postgres
itself, and pg_ctl didn't reused it. That was also the reason of 'write_to_pipe'
existence: if pg_ctl restart's stdout was fed to pipe, communicate() hanged,
reading stdout of postgres infinitely. Now we ride on 'start' routine for
restart and get the same behaviour, including logging.

'inp' param of safe_psql and psql was renamed to input for consistency (we don't
use here standard input() function quite often, right?), test for it was added.