Testgres tests create log dir in exact place
When we do not define TEST_CFG__LOG_DIR it is expected the testgres tests will create a log directory in a root of testgres project folder.

We used config.rootpath for detect this folder in pytest_configure function.

It was occurred that config.rootpath can point to another (unexpected) place.

So we will use exact code to calculate testgres project folder (see TestStartupData.GetRootLogDir) to avid this problem.