[Refactoring] Default port manager functions now use PortManager__Generic and LocalOperations (#251)
* [Refactoring] Default port manager functions now use PortManager__Generic and LocalOperations

This patch deletes a duplication of port manager code.

Now utils.reserve_port and utils.release_port works through _old_port_manager - it is a global instance of PortManager__Generic that uses a global instance of LocalOperations.

This commit is a part of work for #247.

* [BUG FIX] PortManager__ThisHost::__new__ had MT-problem

After MT-lock we must to check __class__.sm_single_instance again.

Refactoring
 - PortManager__ThisHost::__new__ is replaced with an explicit PortManager__ThisHost::get_single_instance()
 - PortManager__ThisHost::__init__ is deleted